previous commit message was wrong - this commit adds a whole bunch of stuff to get a game simulation going.  Board is now filled with legally positioned pieces.  The three pieces that are going to be output are shown as well.
the simulation will fail with a NullPointerException as I haven't guarded against it.  quit by typing q and enter before then or let it fail

this commit removes a currently usless file.  You may want to add it back in the future but you'll have to modify it in a way that makes it work and that might be challenging...