fixed bugs in simulation scoring in parallel collections
this included piece location hash fixes as well as ensuring that scoring occurs when lines
are cleared regardless of which thread is "on point" for scoring that particular piece location hash
fixed strange debugger bug that would cause it to slow down horrendously when evaluating
BullShit.size - which is not something i asked it to do.  Overrode size and provided a value of 0
as though it is an infinite stream, i never check its size
removed maxSimulations parameter - who needs it?
added selftest to ensure parallel simulation scoring actually scores all possible lcoations with a test
that validates this
discovered that i had been scoring multi-line clearing incorrectly.  fixed that - thanks again, Kevin
more logging improvements
added tests for GamePieces and random seeding